Refactor: Split map parsing logic into separate file `parse_map.c`
